requirements-dev.txt
certifi==2019.9.11
cloudinary==1.18.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
django-crispy-forms==1.7.1
django-multiselectfield==0.1.8
gunicorn==19.9.0
mock==3.0.5
Pillow==6.2.1
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7